Style #8 - 테스트 코드 이동
* CustomStreamSupportTest.interleaveTest()
