+++ src/test/java/study/subject/upa/util/DataUtilTest.java
@@ -32,18 +32,4 @@ class DataUtilTest {
         assertThat(flag4).isFalse();
         assertThat(flag5).isFalse();
     }
-
-    @Test
-    public void interleaveTest() throws Exception {
-        // given
-        Stream<String> a = Stream.of("a", "c", "e");
-        Stream<String> b = Stream.of("b", "d", "f");
-
-        // when
-        Stream<String> ab = CustomStreamSupport.interleave(a, b);
-
-        // then
-        String result = ab.collect(Collectors.joining());
-        assertThat(result).isEqualTo("abcdef");
-    }
 }
